% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_enriched_signature_GSEA.R
\name{common_enriched_signature_GSEA}
\alias{common_enriched_signature_GSEA}
\title{common_enriched_signature_GSEA}
\usage{
common_enriched_signature_GSEA(data_GSEA, fdr_threshold = 0.25)
}
\arguments{
\item{data_GSEA}{A named list of dataframes coming from the output of the function read_GSEA_result}

\item{fdr_threshold}{The fdr threshold used to decide which molecular signature is correlated with a differential.}
}
\value{
A vector of class character (empty or not) that contain the name of all molecular signatures that are supposed to be enriched in all expression differential.
}
\description{
Function that allow to identify molecular signatures that are enriched in several differential. Of course, the set of signatures tested should be the #' same for each signature.
}
\examples{

 path_result1 <- GSEAExploratoryTools_example("gsea_report_for_CB_1725354001411.html")
 path_result2 <- GSEAExploratoryTools_example("gsea_report_for_CC_1725353944800.html")
 result1 <- read_GSEA_result_html(path_result1)
 result2 <- read_GSEA_result_html(path_result2) 
 several_signatures <-  list("CB_vs_N" = result1,"CC_vs_N" = result2 )

 common_enriched_signature_GSEA(data_GSEA = several_signatures,fdr_threshold = 0.25)

}
