% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_FDR_qval.R
\name{find_FDR_qval}
\alias{find_FDR_qval}
\title{find_FDR_qval}
\usage{
find_FDR_qval(gsea_report_list, signature)
}
\arguments{
\item{gsea_report_list}{a  R named list that contains one list of 2 dataframes for each comparison (corresponding to the 2 tsv files that give the enrichment in a condition vs another and the reciprocal). The name of each element correspond to the name of the comparison associated to the path.}

\item{signature}{a R character element that contains the name of the signature for which the user wants an FDR-qval value for example "HALLMARK_E2F_TARGETS".}
}
\value{
A R named list. Each element of the R list corresponds to the FDR-qval value associated to a comparison between two conditions.
}
\description{
Function that allows to retrieve FDR-qval value for a specific gene signature in GSEA results directory.
}
\examples{
pathresult1 <- GSEAExploratoryTools_example("GSEA_CB_vs_N")
pathresult2 <- GSEAExploratoryTools_example("GSEA_CC_vs_N")

toimport <- list(
"CB_vs_N" = pathresult1,
"CC_vs_N" = pathresult2
 )
results_tsv <- import_GSEA_tsv_report(toimport) 
 
find_FDR_qval(results_tsv,"KEGG_CELL_CYCLE")
}
