% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAHelpTools_example.R
\name{GSEAExploratoryTools_example}
\alias{GSEAExploratoryTools_example}
\title{gseahelptools_example}
\usage{
GSEAExploratoryTools_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
Get path to gseahelptools examples

gseahelptools comes bundled with some example files in its `inst/extdata`
directory. This function make them easy to access.
}
\examples{

GSEAExploratoryTools_example("gsea_report_for_CB_1725354001411.html")

}
