% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_GSEA_tsv_report.R
\name{import_GSEA_tsv_report}
\alias{import_GSEA_tsv_report}
\title{import_GSEA_tsv_report}
\usage{
import_GSEA_tsv_report(list_path_gsea_output_directory)
}
\arguments{
\item{list_path_gsea_output_directory}{a R list (named R list). Each element of this list is a character chain specifying the absolute path to a GSEA output directory.
Ideally the name of each element of the list gives the comparison associated to the GSEA output directory.}
}
\value{
a R list. Each element of this list is also a list of 2 dataframes (one for the signature enriched in the first condition and 
one for the signature enriched in the second condition)
}
\description{
Function that allows to import the tsv gsea report from several GSEA output directory in a R list (each element of the list corresponds to a specified comparison, that is a specific GSEA output directory).
}
\examples{
pathresult1 <- GSEAExploratoryTools_example("GSEA_CB_vs_N")
pathresult2 <- GSEAExploratoryTools_example("GSEA_CC_vs_N")

toimport <- list(
"CB_vs_N" = pathresult1,
"CC_vs_N" = pathresult2
 )
results_tsv <- import_GSEA_tsv_report(toimport) 
 
results_tsv



}
