% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NES_signature_barplot.R
\name{NES_signature_barplot}
\alias{NES_signature_barplot}
\title{NES_signature_barplot}
\usage{
NES_signature_barplot(signatures, col_signature, gsea_report_tsv_path)
}
\arguments{
\item{signatures}{character vector. Names of the signature for which the user wants a representation of NES.}

\item{col_signature}{named character vector specifying colors associated to each condition. The names of the elements corresponds to the names of comparisons.}

\item{gsea_report_tsv_path}{R named list that gives the path associated of the GSEA results directory for each comparisons. 
The name of each element specifies the comparison associated with the path.}
}
\value{
a ggplot2 representation of the NES for each signature for each comparison.
}
\description{
Function that allows to represent the value of the NES of several signature
}
\examples{

pathresult1 <- GSEAExploratoryTools_example("GSEA_CB_vs_N")
pathresult2 <- GSEAExploratoryTools_example("GSEA_CC_vs_N")

NES_signature_barplot(signatures = c("KEGG_CELL_CYCLE","KEGG_JAK_STAT_SIGNALING_PATHWAY"),
gsea_report_tsv_path = list("GSEA_CB_vs_N" = pathresult1,"GSEA_CC_vs_N" = pathresult2),
col_signature = c("GSEA_CB_vs_N"= "red" ,"GSEA_CC_vs_N" = "blue"))

}
