% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_GSEA_result_html.R
\name{read_GSEA_result_html}
\alias{read_GSEA_result_html}
\title{read_GSEA_result_html}
\usage{
read_GSEA_result_html(path_html)
}
\arguments{
\item{path_html}{R character chain that specify the absolute path to a GSEA results file for a specific comparison, that is the html file 
that gives te top of the signature the most positively correlated to a condition vs another or the opposite depending on the file.}
}
\value{
a R dataframe that gives the top correlated signature with all the associated statistics.
}
\description{
read_GSEA_result_html
}
\examples{
path_gsea_result_html <- GSEAExploratoryTools_example("gsea_report_for_CB_1725354001411.html")

result <- read_GSEA_result_html(path_gsea_result_html)

}
